Create Audiences class.
